move functions from test_utils.py to new file_utils.py: - open_txt() - open_json() - save_json() - open_jsonl()
