+++ gamechangerml/scripts/make_training_data.py
@@ -16,6 +16,7 @@ from gamechangerml.src.model_testing.query_es import *
 from gamechangerml.src.services import ElasticsearchService
 from gamechangerml.src.utilities.text_utils import normalize_query
 from gamechangerml.src.utilities.test_utils import *
+from gamechangerml.src.utilities import open_json, save_json
 from gamechangerml.api.utils.pathselect import get_model_paths
 from gamechangerml.scripts.update_eval_data import make_tiered_eval_data
 from gensim.utils import simple_preprocess
