+++ gamechangerml/scripts/run_evaluation.py
@@ -13,6 +13,7 @@ from gamechangerml.configs import (
     QexpConfig,
 )
 from gamechangerml.src.utilities.test_utils import *
+from gamechangerml.src.utilities import open_json
 import argparse
 import os
 import logging
