+++ gamechangerml/src/model_testing/validation_data.py
@@ -1,11 +1,13 @@
 import logging
 import pandas as pd
 import numpy as np
+import json
 from collections import OrderedDict
 from gamechangerml.src.utilities.text_utils import normalize_answer, normalize_query, filter_title_queries
 from gamechangerml.src.utilities.test_utils import *
 from gamechangerml.configs import ValidationConfig
-from gamechangerml.src.utilities.test_utils import filter_date_range, open_txt
+from gamechangerml.src.utilities.test_utils import filter_date_range
+from gamechangerml.src.utilities import open_txt, open_json, open_jsonl
 from gamechangerml import REPO_PATH
 
 CORPUS_DIR = os.path.join(REPO_PATH, "gamechangerml", "corpus")
